+++ docs/docs_streams-6.py
@@ -1,0 +1,3 @@
+import convis
+n = convis.streams.PseudoNMNIST()
+convis.plot(n.get(200))